:recycle: api: move answer-repository to shared
