+++ api/lib/domain/events/index.js
@@ -7,7 +7,7 @@ import _ from 'lodash';
 import perf_hooks from 'perf_hooks';
 import * as eventBusBuilder from '../../infrastructure/events/EventBusBuilder.js';
 
-import * as answerRepository from '../../../src/evaluation/infrastructure/repositories/answer-repository.js';
+import * as answerRepository from '../../../src/shared/infrastructure/repositories/answer-repository.js';
 import * as authenticationMethodRepository from '../../infrastructure/repositories/authentication-method-repository.js';
 import * as assessmentRepository from '../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import * as assessmentResultRepository from '../../../src/shared/infrastructure/repositories/assessment-result-repository.js';
