+++ api/lib/domain/services/certification-challenges-service.js
@@ -12,7 +12,7 @@ import { KnowledgeElement } from '../models/KnowledgeElement.js';
 import { Challenge } from '../../../src/shared/domain/models/Challenge.js';
 
 import * as challengeRepository from '../../../src/shared/infrastructure/repositories/challenge-repository.js';
-import * as answerRepository from '../../../src/evaluation/infrastructure/repositories/answer-repository.js';
+import * as answerRepository from '../../../src/shared/infrastructure/repositories/answer-repository.js';
 import * as knowledgeElementRepository from '../../infrastructure/repositories/knowledge-element-repository.js';
 import * as learningContentRepository from '../../infrastructure/repositories/learning-content-repository.js';
 import * as certifiableProfileForLearningContentRepository from '../../infrastructure/repositories/certifiable-profile-for-learning-content-repository.js';
