+++ api/src/shared/infrastructure/repositories/answer-repository.js
@@ -2,8 +2,8 @@ import _ from 'lodash';
 import jsYaml from 'js-yaml';
 import { knex } from '../../../../db/knex-database-connection.js';
 import { ChallengeAlreadyAnsweredError, NotFoundError } from '../../../../lib/domain/errors.js';
-import { Answer } from '../../domain/models/Answer.js';
-import * as answerStatusDatabaseAdapter from '../../../../src/shared/infrastructure/adapters/answer-status-database-adapter.js';
+import { Answer } from '../../../evaluation/domain/models/Answer.js';
+import * as answerStatusDatabaseAdapter from '../adapters/answer-status-database-adapter.js';
 
 function _adaptAnswerToDb(answer) {
   return {
